refactor: 환경변수 설정
