+++ src/App.tsx
@@ -1,13 +1,13 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import { QueryClient, QueryClientProvider } from 'react-query'
 import { ReactQueryDevtools } from 'react-query/devtools'
-import MainPage from './presentation/pages/MainPage'
-import PostDetailPage from './presentation/pages/Post/PostDetailPage'
-import PostPage from './presentation/pages/Post/PostPage'
-import HeaderComponent from './presentation/components/HeaderComponent'
-import Kakao from './presentation/components/account/Kakao'
-import Google from './presentation/components/account/Google'
-import SignUpPage from './presentation/pages/SignUpPage'
+import MainPage from 'pages/MainPage'
+import PostDetailPage from 'pages/Post/PostDetailPage'
+import PostPage from 'pages/Post/PostPage'
+import HeaderComponent from 'components/HeaderComponent'
+import Kakao from './components/account/Kakao'
+import Google from './components/account/Google'
+import SignUpPage from './pages/SignUpPage'
 
 const queryClient = new QueryClient()
