+++ src/components/account/Login.tsx
@@ -6,12 +6,12 @@ import { useMutation } from 'react-query'
 import useServiceManager from 'src/hooks/useServiceManager'
 import styled from 'styled-components'
 import { googleURL, kakaoURL } from 'src/utils/url'
+import kakaoImg from 'img/kakaoLogo.svg'
+import googleImg from 'img/googleLogo.svg'
+import showPwImg from 'img/showPW.svg'
+import hidePwImg from 'img/hidePW.svg'
+import { AccountInfo } from 'types/account'
 import { ErrorEmail, ErrorPassword } from '../Error'
-import { AccountInfo } from '../../../types/account'
-import kakaoImg from '../../../assets/images/kakaoLogo.svg'
-import googleImg from '../../../assets/images/googleLogo.svg'
-import showPwImg from '../../../assets/images/showPW.svg'
-import hidePwImg from '../../../assets/images/hidePW.svg'
 import Button from '../AccountButton'
 
 const {
